osif: Adapt ZBOSS thread routine in NCP mode
There is an implementation of zboss_app_main() in the ZBOSS library.
The modification is to use it in NCP mode instead of the previous
stack thread procedure as the latter has become incompatible.

The mentioned implementation has been slightly adapted to nRF5,
hence the update of libncp-dev as well.

Signed-off-by: Adam Zelik <adam.zelik@nordicsemi.no>